Styles improved
